chore: updated project with hcms core

+++ src/utils/dateFormatter.ts
@@ -13,7 +13,7 @@
  */
 export const formatDate = (isoDateString: string): string => {
   const date = new Date(isoDateString)
-  const month = date.toLocaleString('default', { month: 'long' })
+  const month = date.toLocaleString('default', { month: 'short' })
   const day = date.getDate()
   const year = date.getFullYear()
   return `${day} ${month}, ${year}`
